Rework the way applet configurations are handled
Instead of having a ScrollablePage with a loader push the applet's
config directly onto the pagestack.

This assumes that the config's root item is one of the KCM QML
components (SimpleKCM, ScrollViewKCM, GridViewKCM).

This gives the applets more flexibility and control over the behavior,
like when and how to use a scrollview

Implements https://invent.kde.org/plasma/plasma-desktop/-/issues/99