containment/panel: fix some incorrectly ported plasmoid properties
In this case, lowercase-p "plasmoid" was actually a local variable, so
porting it to the uppercase version caused the variable to be ignored
and instead we were referencing the data from the capital-P Plasmoid in
the config poverlay--that is to say, from the panel itself.

Fix this and also rename the variable for clarity.

BUG: 473844