kcms/access Fix unintentional config changes
Qml spinboxes have a default range of [0, 99] which is less than
the default values for SlowKeysDelay and BounceKeysDelay set in
kcmaccessibilitykeyboardfilters.kcfg.

This causes the SpinBox's value to be set to it's upper bound
instead. In combination with the "onValueChanged" handler the
config gets updated without user interaction.

This is unintentional behaviour as the user never initiated
this change. The config should only be updated when the user
initiates the change. "onValueModified" does just that.