Kickoff: add option to move sidebar to the right side of the popup
Always placing the category sidebar on the left made it an obstacle to
reaching the favorites when Kickoff was located at the left edge of the
screen. The direct path to select an item from the Favorites list
required the user to move the mouse across the sidebar area, and if the
journey across sidebar items took longer than half a second or
curved/veered in a way that the triangle filter doesn't expect, it would
be interpreted as a hovering sidebar selection, removing the favorites
from view and disorienting the user. Since humans don't always move mice
in quick, perfectly linear motions, this frustrated more than a few
people rather often.

It was especially problematic when Kickoff was at the bottom left corner
of the screen, because the sidebar's Favorites category item (which is
always at the top of the list) was then as far away from the mouse as
possible, adding insult to injury by making it extra inconvenient to
find and click whenever the user had to restore the favorites to view.

Then, after restoring the favorites to view, the original experience
could easily repeat as the user once again moved the mouse from the
sidebar's Favorites category item, across the sidebar, toward their
target in the Favorites list.

Solution:

This patch adds a config option to place the sidebar on the right,
so people can have their favorites near the mouse when it opens.

Layout mirroring is respected, so Tab key navigation should work just as
well in locales with right-to-left Tab key navigation as it does with
left-to-right.

BUG: 443082
FIXED-IN: 6.0