kcms/componentchooser: Stop writing legacy default browser key
The x-scheme-handler mime types are the "modern" way of setting the
default browser. However, we were still writing the kdeglobals key,
presumably for backward compatibility. This potentially causes issues
because things may read that key instead of the scheme handler.
Moreover, it is yet another key that can get out of sync.