kcms/keyboard: Add shortcut to toggle last-used keyboard layout
Often when documents are edited--especially when users use non
Latin-based layouts--it's useful to quickly switch between two
layouts. Users whohave more than 2 keyboard layouts have to use
custom shortcuts for each layout, or have to toggle through all
layouts to reach the desired one.

This commit adds a new option in the keyboard layout settings
settings which allows setting up a new keyboard shortcut that
will switch between the current and the last-used layout,
independently of their positions in the layouts list or the
number of layouts configured.