desktoppackage: Port table heading model to ListModel
HorizontalHeaderView, being a descendant of TableView, does not like
being assigned a JavaScript array of strings as a model. In some cases
it may crash (as in QTBUG-117917), in other cases like ours it may not
display any text with obscure "Unable to assign [undefined] to QString"
error message for its `text: model[controlRoot.textRole]` binding.