Port highlight effects to use PlasmaExtras.Highlight and unify behavior
Porting to this component makes the visual styling consistent and lets
us change it in the future--if need be--from only one place. Some things
that were already using it did not specify enough conditions to have it
cycle through all of its various states, and that's been corrected here
too.

Part of https://invent.kde.org/plasma/plasma-workspace/-/issues/97.