Hide lauchers when not "separate launchers"
After https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/1744
launchers are always present in the traditional task manager no matter
what. also, the separateLaunchers option doesn't make much sense putting
launcher icons besides their running task.

In a discussion in that merge requests was agreed that this option
should intead hide launchers in TTM to keep a behavior similar to the
previous, but was merged without it, so this implements that.

Not being possible to hide launchers anymore is kinda a problem for ttm
as its main feature is to keep window titles as readable as possible,
and if one has a large amount of launchers, they go to bite in the space
for the labels.

I do get the reasoning for having this behavior as default as it gives
always an immediate feedback to the "add launcher" actions, but it just
doesn't work for me.