kcms: Add new Game Controller KCM
This is based on QtQuick and Kirigami, and features a much cleaner and
modern codebase. It's functionally equivalent to the old KCM, save for
a tracing and calibration feature. It's designed to be a drop-in
replacement and a base for future improvements and features.

Internally, the KCM name is updated to "kcm_gamecontroller" in many
places where "kcm_joystick" was used.