Implement UI to enable "Dodge Windows" mode for the panel
Depends on https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/3508.

This adds the configuration UI and tweaks the limit for deciding
whether a window is touching to be 12px instead of 6px. After testing
various values for it, I found 12px is the smallest that prevents
glitches with window snapping.

CCBUG: 349785

![Screenshot_20231110_010435](/uploads/e042df56fdf596f3c2be0c70e52a31af/Screenshot_20231110_010435.png)