PanelConfiguration: fix clicking position preview and using arrow keys
When you click the "Set Position…" button, the buttons for changing the
position appear on screen edges, and you can also use the arrow keys to
change the position.

However when you click on the preview image above the button, the
buttons for changing position appear on screen, but the arrow key
shortsuts do not work. This is because the key handlers only live on the
button, which does not have focus after you click on the preview. image.

To fix this, force focus to the button when clicking on the preview. The
preview doesn't accept focus anyway so this shouldn't be a problem.


(cherry picked from commit 0fbc4a344189fc8542170c61d0f76f36d0beaf3c)