Remove old panel focus workaround
Simplify panel focus from shortcut.
The logic was done to fix bug 472909 (focus not getting back to
proper window after opening kickoff with shortcut), replace it
with a better fix in plasma-workspace.

See https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/3868

That logic caused a breakage in the shortcut to give panel focus,
as it became necessary to press it twice to give panel focus
for the first time after plasma startup