Folder View: remove "Refresh" action from context menu
This item exists as a workaround for when the contents of the directory
have been updated, but failed to visually refresh as expected. However
today this is a relatively rare occurrence, especially with
kde-inotify-survey prompting people to increase their number of file
watches which generally fixes the issue.

The issue that people do still suffer from is when items are missing due
to glitches in the model or the display; we have a number of such bug
reports. This action does nothing to resolve that issue, misleading and
frustrating users.

The item also consumes space in the menu, whose length we've previously
received criticism for.

F5 still refreshes the view if you really do need to do that for some
reason.