kcms:keys: fix changing command text or path
The code was trying to change an item by its ID that is dynamically
instantated within a delegate; that won't work.

Instead we need to change a top-level property on the delegate itself
and bubble that change down to its contentItem.

BUG: 481825
FIXED-IN: 6.0.1


(cherry picked from commit e9600f773fdbcba72778a1d3f36d25290565bbcf)