applets/kicker: add power, session, calc and unit converter runners
There is a limitation on the number of runners we can add here since
every runner adds a column, making the popup wider and wider. However
these runners return results that are infrequently matched by accident,
limiting the frequency with which they will appear and clutter up the
UI.

As such, I think adding them does not result in a worse UI and we can
safely do it even before we fix it to not grow wider with every new
runner's search results.

BUG: 375857
BUG: 480692
FIXED-IN: 6.1