Rework logic to know who is the first visible results view
We have a horizontal layout of runner results that separate the results
by category.

In order to show enough space for the text field and to disable the
first separator line, it assumed the item with index===0 was the first
element.

However if there aren't results which are applications, this assumption
would break and the layout was wrong, so check that this column is
actually the first visible child.

BUG: 482736
FIXED-IN: 6.0.2


(cherry picked from commit e0582b167be904abedb2332ad3f459eab52e1d4e)

f5237fb8 Rework logic to know who is the first visible results view
db6d7462 fix resizing