Shorten long action names in desktop toolbox and context menu
A lot of these menu items are really long, unnecessarily so. For the
ones that take the user to another location rather than performing an
action, we can make the menu item's text simply echo the name of the
place they'll be taken to. This is what we already do for some other
similar menu items that take you to another location, like "Properties".

BUG: 481756
FIXED-IN: 6.1