Folder View: fix label rendering for popup and widget views
The shadow only needs to be displayed on the desktop, in containment
mode. Previously we were checking only for whhether or not the widget
was in list mode, which missed the case of grid mode for standalone
widgets and folder popups on the desktop.

BUG: 484163
FIXED-IN: 6.0.3


(cherry picked from commit d5878b2d173e21a6b85b5effda9cdb2819686559)