applets/taskmanager: Reset winId to undefined in ToolTipInstance
The ToolTipInstance is kept persistently. When a task is hovered or
clicked, the bindings will be updated and the winId can be reset to 0.

ScreencastingRequest views "0" as a valid win id and it's going to
ask kwin to create a stream for a window with such an id. kwin will
likely respond with an error.

This change ensures that the winId is reset to "undefined", so the
ScreencastingRequest can properly reset the node id.