applets/kickoff: make section header text more reasonably sized
Currently Kickoff's section headers are using Labels with a custom text
size, and it's much too big. Replacing that with a Kirigami.Heading is
the semantically correct thing to do, and it looks better too.

This was always an issue, but it became much more obvious with the
recent change to show section headers in search results lists.