kcms/mouse: also filter mice that say they're keyboards
When enumerating devices to identify mice, the KCM already filters out
pointer devices that say they're touchpads. But it doesn't do so for
devices that say they're keyboards.

This is weird, but apparently a lot of gaming mice report themselves as
multiple devices with one being a keyboard that has only a couple of
buttons on it.

Let's filter these out so that we only see the mouse representation of
these devices in the KCM.

BUG: 465957
FIXED-IN: 6.1