applets/taskmanager: show text more often for traditional Task Manager
The minimum area for text was using a PlasmaCore unit that got ported to
Kirigami.Units.iconSizes.sizeForLabels for Plasma 6. This has different
semantics, resulting in the minimum size to show text increasing hugely.
As a result many people stopped seeing text in their task managers with
medium-thickness vertical panels crowded horizontal panels.

The more semantically correct iunit to port to is GridUnit. It's not
perfect though since it's the font height, not width. But it's the best
we've got.

In my testing, 1.5 of them  works perfectly for both vertical and
horizontal panels to only show text when there's at least one character
that won't be elided. This isn't a lot of text, but pepole use a
traditional Task Manager because they want to see text, so it at least
better matches their expectation than the status quo.

BUG: 483070
FIXED-IN: 6.0.4


(cherry picked from commit 8608d7034199e9d19f3af3837d4b2f6baa8c8f4c)