kcms/mouse: Allow configuring bluetooth mice
Bluetooth mouse devices are usually reported both as keyboards and
pointers. Because of that, it's not possible to configure such devices
on wayland at the moment because of the keyboard check.

It's weird to display bluetooth keyboards too, but it's far worse not
to be able to configure mouse devices.