applets/showActivityManager: Fix and improve layout
Fixes an issue where the showActivityManager applet was too small and
caused some applets on the panel to overlap.

Also fixes two other issues found while testing the change:

1. There was an extra small spacing on the right of the applet when not
   displaying the activity name.
2. The activity name could go past the widget when placed on the desktop.

One other pre-existing problem that I couldn't solve is that when you place
a new showActivityManager widget on the desktop, the initial width isn't
big enough to contain the icon and text.

BUG: 482098
FIXED-IN: 6.0.5