PanelConfiguration: make it more obvious how you close it
Right now, there's no clear way to close the panel configuration dialog.
It lacks both an OK or Done button like you'd see in a dialog, and also
a close button in the fake titlebar like you'd see in a real window.
The dialog also doesn't disappear on focus loss like the old UI did.

This isn't ideal. Let's address it by adding a "Done" button to the
bottom-trailing position, where people generally expect to see buttons
to close dialogs.

I chose this over a close button in the fake titlebar because a close
button (especially a red one, as the Breeze icon is) might be seen more
negatively, i.e. "throw away settings". Whereas a "Done" button with
a checkmark is clearly an affirmative action that will accept and save
any changed settings.

To make room, the Delete Panel button is moved to the other side of the
footer, and the shortcut UI is added into the body of the dialog, where
the thickness control lives.

---

Thanks to @eatsu for starting this off in https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/2021.

This is an intentionally minimal change to fix a targeted UX issue; let's try to avoid redesigning the entire dialog or else we'll go in circles like we did there.

Co-authored-by: Taro Tanaka <mkrmdk@gmail.com>

BUG: 483299
BUG: 486703
FIXED-IN: 6.1

@teams/vdg @teams/usability 

| Before | After |
| ------ | ----- |
| ![Before](/uploads/e74e84963544a4d4a5e35ed4245ed951/Before.jpeg) | ![After](/uploads/17e217a34b5ff5809491764afcb5ef32/After.jpeg) |