panel: Fix an error when dragging a piece of text over a panel
Usually, child is an instance of a Loader with a `property Item applet`
where applet usually an instance of PlasmaQuick::AppletQuickItem (or
rather its subclass PlasmoidItem to be precise). However, applet could
also be an instance of a dndSpacer placeholder, which is just an Item.

Fixes the following error when dragging a piece of text over a panel:

file:///usr/share/plasma/plasmoids/org.kde.panel/contents/ui/LayoutManager.js:91:
TypeError: Cannot read property 'id' of undefined