foldermodel.cpp: add context to menu connection
Menu connection was missing "this" as context, so it would be called
even if "this" was not there anymore.

The bug I've linked before is very difficult to reproduce, it seems to 
be a race condition that triggers rarely for some, more often for others.
I have only gotten it happen once, but the backtrace indicates this seems
to be the problem.

CCBUG:463847