kcms: Reorganize settings in mouse and touchpad KCMs
Mouse KCM: Switch order for the "Scrolling speed" and the "Scrolling"
options. Reason for this is that the "Pointer speed" is the first option
for the pointer settings so it makes sense to have "Scrolling speed" as
the first option for the Scrolling settings.

Touchpad KCM: Move the "Scrolling options" up to be under the "Pointer
acceleration" option so it has the same order as the Mouse KCM.