kcms/workspaceoptions: Add smooth scroll setting
Most software that implements smooth scrolling itself also has an option
to enable/disable it based on the user's preference, so I think it's a
good idea for KDE to have this option as well. I previously discussed
this with some of the Kirigami developers (as most of the KDE software
that has smooth scrolling is built with Kirigami) and they said they'd
prefer to make it a system-wide setting rather than something like a
Kirigami-specific environment variable.