kcms: landingpage: Use raised styling for "most used KCMs" buttons
This KCM has other buttons that take you different KCMs, and they're
raised. But the "most used" buttons are not, they're flat. This is
inconsistent, so make them raised.

I'm not 100% in love with the visual result, but I'm not sure how to
improve upon it without a larger redesign of how we present most used
KCMs here.

BUG: 487648
FIXED-IN: 6.2