applets/pager: Use NumberAnimation for opacity animation
The animation driver is shared among windows, unless a QQuickRenderControl
object is created for a particular window. In addition to that, the
QSGThreadedRenderLoop doesn't have a notion of window affinity for
animations.

So, if there are two windows and one of them is hidden and has an item
whose opacity is animated using OpacityAnimator, then the threaded render
loop will force repaints for the other window even though it has no
animations or nothing has updated in it. It is a design flaw in QtQuick.

To remedy that issue, this change ports the pager to NumberAnimation for
the opacity animation.

CCBUG: 487895