applets/pager: Remove animation on window size
The window miniatures are typically small to the point where it is
not noticeable whether they are animated unless somebody looks at
them really close. On the other hand, since the size of the miniatures
is animated, it can make the panel force screen recomposition even
though nothing changes on the screen except the window miniature.

Also, it makes the visuals of the pager more consistent because the
position of the miniatures is not animated. And there are good reasons
to avoid animating them because that will introduce a new set of issues
since there are applications that update their position in
short bursts of time, usually due to the client side session restoration,
which can result in the window miniatures jumping unexpectedly, etc.