FolderView Positioner: fix dragging files across screens.
f0ede2d83731952012be0d43c77916d18617edb2 fixed a bug in which items
would disappear when dragging between screens. The rationale in the
description was correct. The fix there was to queue the connection of
rowsInserted so that we don't emit move signals before Positioner proxy
model has processed any pending rowsInserted signals.

This was incorrectly simplified in
99a3597285a92531efb09492421fac0a6a2e82dc . The important part isn't to
process it after FolderModel gets a source model set, but to process it
after Positioner gets the source model, which FolderView doesn't know
about and the original bug introduced.

Rather than a straight revert the code is adjusted to emit the move
signal delayed. This is safer than delaying the processing of a signal
with index arguments.

BUG:481254
BUG:466337


(cherry picked from commit 4da2346e42302349422cf7c45e64e838fb4d630f)

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>