Revert "Folder View: with menu representation, always open with a single click"
This reverts commit d5abf12d6661f2a8f8449d9abe88468114a73dca.

The change to always open items in single-click when using list mode
was added by me in
https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/1467 with
two +1s and no discussion. But in retrospect, I think it was not the
right change; the list-view popup may look like a menu and kind of acts
like a menu, but it's not a menu--it's a file view in list mode. We
don't force single-click for other such views in Dolphin, file dialogs,
Kate, or other place where there are file views.

BUG: 484026
CCBUG: 420261
FIXED-IN: 6.2