kcms/baloo: move status to footer
The status information includes the filename, which is of variable
length. Unless this is on the very bottom of the page, it'll cause all
the UI elements below it to jump around as it changes; not ideal.

Move it to the footer, where it assumes the comfortable and common role
of a status bar.

BUG: 487898
FIXED-IN: 6.2