Refactor drag and drop behaviour of applets from panel to other containments
Right now it's impossible to drag and drop applets from the panel to elsewhere, only within the panel. If you try, the applets will be added to the desktop right next to the panel, and you'll have to start dragging them again to move them around or to another panel. This implements a new mode to drag and drop applets from the panel to other containments.

This MR makes it possible to just drag and drop items outside the panel; when you do so, a "real" drag and drop will start that allows you to move that applet on the desktop or to other panels. You can also put the applet back to the origin panel.

Requires https://invent.kde.org/plasma/libplasma/-/merge_requests/1142


Co-authored by Marco Martin


BUG:487825