kcms/access: fix list item highlighting
This is kind of a manual list view, so the items are responsible for
setting their own highlighted state. Without that, the active one won't
be highlighted, and there won't be a difference between the hover and
highlight effects.

BUG: 489206
FIXED-IN: 6.1.2


(cherry picked from commit 7aeebe51ade45b0a924252c591963d05ba205b45)

Co-authored-by: Nate Graham <nate@kde.org>