applets/taskmanager: remove inline grouping width special case
There is no indication that using inline grouping (grouping as multiple
buttons, with "Combine as single button" turned off) would change
anything about the width of tasks items, so it should not be tied
together.

The special cased max width is also too small, with eg. the label for
the default page in System Settings already clipping.

The special case was added as part of https://phabricator.kde.org/D1882
just to match the behavior of Netrunner's Expanding Icons Task Manager
fork from back then, which the feature was a port of.

BUG: 489375
FIXED-IN: 6.2.0