kcms/access: fix accepted file types and description
We also support .oga files here, but the dialog wasn't set up for it.
However even if it was, you wouldn't know because the user-facing
description of the accepted file types didn't clarify which exact file
types are supported. Because we only support certain file types here, we
need to communicate that to the user too.