kcms/tablet: Use a QML MouseArea to accept tablet events in the tester
We were using the same method in other parts of the KCM to monitor
tablet events. This is only really useful passively, as it does not
intercept said events and pass them right along.

This has an accidental effect, as is the case with the new tablet tester
tool. In Qt 6.8, tablet tools can now drag windows which hid a bug here.
Since we passively monitored tablet events, it will send  regular
tablet events through and onto the window the tester sits on. This makes
it drag while trying to draw.

A simple way to fix this is by putting an active input handler in the
drawing area to catch said events.

BUG: 489874