Don’t extract QML text as Java strings
Extracting the strings as Java strings causes the string ‘%1%’
to be marked as ‘java-printf-format’ in the translation file.
The result is that it can’t be properly translated in all
languages, e.g., into ‘%1 %’ for languages which require a
space in front of the percentage sign.

See https://mail.kde.org/pipermail/kde-i18n-doc/2021-March/000450.html
and
https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/711
for more details.