kcms/mouse: follow new HIG guidance for header button & fix translation context
The guidance provided in the KDE HIG for button and menu length was
recently updated to provide an explicit length target of three words,
with a total width not longer than "Configure Keyboard Shortcuts…".

The header button that opens the Mouse Button configuration page does
not conform to this yet. This change updates the label from "Re-bind
Additional Mouse Buttons…" to "Configure Extra Buttons…" and modifies
the title of the resulting page to reflect this. This commit also
contains a tiny fix to the translation context of the title of that
page.