kcms/keys: prepare setting Action Accesible.names
The Kirigami.Actions often omit the object in their text and only
contain the verb. This can be hard to understand without being able to
look at the context. Unfortunately, setting Accessible.name on Actions
does not work correctly with Qt 6.7.

This change prepares for the switch to Qt 6.8 by setting up the required
properties, but commented out. Each comment is clearly marked with a
TODO and a indication that these properties can be enabled after 6.8 is
the required Qt version. These lines are untested and should have their
functionality tested when they are ready to be enabled.