applets/taskmanager: Skip querying KActivitiesStats if recentdocs is disabled
When recent document tracking is disabled, the kactivitymanagerd Stats
plug-in stops recording them in the SQLite database.

However, ResultSet will still query the database. If the user's home directory
is on an NFS share, this can lead to a noticeable delay when opening the
task manager context menu.

For this setup, it's best to disable recent document tracking as such
entirely, and skip querying the ResultSet in the first place.

It would also match user expectation to stop seeing recent documents
as soon as the feature is disabled.

Signed-off-by: Sebastian Stark <sebastian.stark@advantest.com>
(cherry picked from commit 581e6b76e15d70cb768002a4dd2540f333dc2929)