applets/taskmanager: Fix applet's layout size with multiple rows
The property `Layout.maximumWidth` of the `taskList` didn't take
into account the number of rows. This was compensated later when
calcualting its actual width, but not for the applet's preferred
size, resulting in extra demanded space.

To fix it, move the influence of the rows directly into the property.

This was only noticeable with the settings `forceStripes==true`,
and `fill==false` because in any other combination, either the space
is wanted, or the tasks fills the available space before creating
a new row.

Due to symmetry, the same is done on vertical panels for
'Layout.maximumHeight` and the number of columns.

BUG: 490319
FIXED-IN: 6.1.4


(cherry picked from commit 71ddcebf3deb5bfa921bbaa59045397cf4d818ff)