kcms/tablet: Store buttons to allow switching between connected pads
Right now there's a limitation in the KCM (that will still be there
when the combined model is added) that you cannot switch between pads.
This is because we only get the button count temporarily, when the pad
appears to our Wayland seat.

There's nowhere where we previously stored this button count, so it
wasn't possible to switch between devices. Now, the button count is
stored in TabletEvents and refreshed when the padDevice changes or the
devices are initially loaded. Everything else is functionally the same.