applets/taskmanager: fix size of textual list popup
The contextual list popup for grouped tasks is restricted to a maximum
size of (screen width / 3) × (screen height / 2). For both, the
denominator is additionaly scaled by the devicePixelRatio. This scaling
does not really make sense, as the values are already in logical pixels,
so the only thing it does is that users with a scaling factor over 100%
get a smaller popup (in terms of actual area) that has space for fewer
items before scrolling becomes necessary. This is particularly
pronounced on Wayland, where the devicePixelRatio seems to be rounded up
to the next integer currently, and even small increases in scaling lead
to a huge reduction in dialog capacity.

This change removes the additonal scaling; the maximum dialog size is
now always a third of the actual screen width and half the actual screen
height.


(cherry picked from commit a0e3ecff776f5d38c511e44039618b3dc8d5c6f2)

Co-authored-by: Christoph Wolk <christoph.wolk@gmx.de>