Folder View: hide existing label while renaming
If we don't, it may peek out from behind the text field with the new
name if the user chooses a shorter name, and that looks weird.

This worked in the past but broke with the port to ShadowedLabel, which
disabled the shadow while renaming when it should have hidden itself
completely.

BUG: 482802
FIXED-IN: 6.1.4


(cherry picked from commit c1ec86898f49d627d59e9afe5fa14efc49c3f949)

96fc78bf Folder View: hide existing label while renaming

Co-authored-by: Nate Graham <nate@kde.org>