Use simple toggle to hide/show clock on lock screen
Part of https://invent.kde.org/plasma/plasma-desktop/-/issues/133.

This replaces the existing option "Clock: Keep visible when
unlocking prompt disappears" with a simple toggle that controls
whether the clock is always or never shown on the lock screen.

From the discussion when the old option was implemented,
(https://phabricator.kde.org/D24893) the use case for the old
option was for users who want an unobstructed view of the lock
screen background. This use case can still be satisfied by the new
option, and perhaps the simplicity of the new option will make it
easier to discover, as the old description in the UI was rather
awkward.

The difference with the new option is that it does not
matter whether the user is currently typing their password. With
the old option, the clock would still be shown when the user moved
the mouse or started typing, which didn't give users a choice to
not see a clock at all.

FEATURE: 472587
FIXED-IN: 6.2.0