Make label of unsupported tablet orientation change clearer
In https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/2400 , the
greyed-out combo-box label of an unsupported change of tablet orientation was
changed from "Default" to "Unsupported".

That sparked the question in that MR if this could be misconstrued as
"Unsupported ... by unfinished Plasma!!"  instead of the correct "Unsupported
by hardware" by a non-technical user. While the new label "Not Applicable"
might not be the optimum label as well, it was regarded as "an improvement" in
regards to "Unsupported" in the base PR linked above this was discussed in.