Treat containment config pages like applet pages in containment config
4afadec2053ee8a44ecaf4124151d78f6c309872 changed the config dialog so that
applet config pages are expected to bring their own scroll view.

However because ConfigurationContainmentAppearance.qml is complicated this wasn't
applied to containments. For this reason containment config pages get a scroll view
provided.

This introduces a problem for folder view, which can act both as an applet and containment.
Depending on which mode it works in the config pages get a scrollview or not.

To address this treat containment-specific config pages like applets, i.e. expect them
to provide a scroll view.

Only supply a scroll view for ConfigurationContainmentAppearance.qml and ConfigurationContainmentActions.qml
which are common to all containments and don't provide a scroll view.

This means that containment config pages need to be adjusted