kcm_keys: fix importing custom commands
This adds new handling for Custom Commands, such as exporting
them as their own group.

It also adds a signal that's emitted when an application is
added to the list. When the signal is emitted, we re-import
our config file to add the missing data. This is because
otherwise we'll face asynchronicity issues.

Most of the work was done by David Redondo already, I just
applied his patch and modified it.

BUG: 488997
FIXED-IN: 6.2.0