desktoppackage: make config sidebar pages PageTabs
In configuration windows, a sidebar allows the user to select various
subpages including the configuration pages, a keyboard shortcut, and the
About page. The Accessible.role of the sidebar and entries is set to
MenuBar and MenuItem, which is misleading as this does not function like
a menu. Qt has a role that fits much better, namely PageTabList and
PageTab "that the user can select to switch to a different page in a
dialog." (QAccessible class).

This change sets the roles of the sidebar and entries to PageTabList and
PageTab.