desktoppackage: make Apply and Cancel tab-accessible
The OK button in the footer is set to pass focus to the sidebar on tab
press, which means that Apply and Cancel cannot be reached with tab
navigation, or any keyboard-based method except accelerators.

This change moves the point where focus is passed back to the sidebar on
tab press to the final button, Cancel, making it and Apply reachable
with keyboard navigation.