kcms/mouse,touchpad: use hyphen for left-handed mode
The common spelling of left-handed in (American) English is with a
hyphen; this form is listed in dictionaries like M-W and is also the
most common in technical contexts for input device modes. The mouse and
touchpad kcms, hover, spell it "Left handed mode".

This change corrects this to the standard spelling "left-handed mode".