applets/showdesktop: limit Minimize All by desktop/activity
The Minimize All widget currently minimizes all windows on all virtual
desktops, which is likely not what the user intended. The same, perhaps
even more counterintuitively, is true for activities.

I have changed this so that the widget now tracks the active state and
the lists of windows that it minimized (if any) separately per
activity/desktop combination. A user can now press Minimize All, do all
sorts of other things on other activities/desktops, come back to the
original desktop in the original activity, press the button again, and
have their windows restored. I think this is a significant QoL upgrade
over the existing version, where restoring will not work if the user
activated any window, even if it was on another activity or desktop. 

This is a change in default behavior with no configuration option. I think 
it's enough to do the right thing and not clutter up the UI with irrelevant
options; see also comment 1 in BUG: 475507. If desired, I can make the
behavior optional later.


FEATURE: 475507
FIXED-IN: 6.2.0