Fix UI of Accessibility settings
This fixes certain UI issues with the Accessibility KCM.

This includes:

* minor UI changes
* fixing inaccuracies
* better explaining the functionality to the user with labels and ContextualHelpButtons
* renaming Activation Gestures to Activation Shortcuts

This does not touch the underlying C++ business logic, and it doesn't do naming changes there.

This makes it so users don't have to resort to documentation or web search to understand the features available to them.

The following are still necessary:

* update the docs to the new UI/terminology
* make the KCM fully accessible
* make the underlying code use consistent terminology
* UI code redesign