kcms/access: Move the Color Blindness Correction settings page here
This makes far more sense than having it in Desktop Effects, which is
essentially an implementation detail which should be moved away from,
especially for accessibility components.

This also slightly reorganizes the page to make more sense within the
context of the Accessibility KCM. It also adds some slider labels and
sets a minimum value higher than 0, because for some reason the KCM
goes back to full intensity when set to 0, but not in the Desktop
Effect configuration. Being able to set it to 0 is redundant anyway
now that there's an enablement checkbox.

See also https://invent.kde.org/plasma/kwin/-/merge_requests/6317
which sets the corresponding Desktop Effect as internal.

FEATURE: 488771
FIXED-IN: 6.2.0