kcms/desktopppaths: use "folder" and update tooltips
The FormData.label for the path entry fields currently calls the XDG
user directories "paths", which seems unnecessarily technical and does
not match the tooltip shown when hovering over the fields, which calls
them "folders". In addition, the tooltips/Accessible.descriptions are
sometimes overly verbose or incorrect (such as stating that changing the
Desktop folder path will also move the files there, which it does not).

This change makes the labels use the term "folder", yielding simpler
labels such as "Music folder" instead of "Music path", makes the
tooltips more compact and informative, and adds translation contexts.