applets/showActivityManager: rename to full wording
This applet is currently called "Activities" in the UI. This is
confusing in multiple ways - it is not the actual activity feature, but
has the same name, its right-click menu shows "Configure Activities" but
does not actually allow the user to configure activities (as opposed to
the Activity Bar applet, where "Configure Activities" opens the activity
kcm, and it makes it harder to later add support for opening the kcm
from the right-click menu as the obvious name for it is already taken.

This change modifies the display name to the somewhat verbose "Show
Activity Manager", already the internal name of the applet. "Activity
Manager" would be another option, but still has the problem of having a
context menu entry called "Configure Activity Manager" that does not
allow actually configuring the Activity Manager.