applets/pager: add activity kcm to context menu
The virtual desktop version of the pager contains a context menu entry
that allows directly accessing the corresponding kcm, as does the
activity bar widget. The activity version of the pager, however, does
not have easy access to the activity kcm.

This change adds such an entry to the context menu and adds translation
context. It also removes the ellipsis from the "Show Activity Manager"
menu entry, as the activity manager is shown without further user input.