Desktop: Fix up sideBarOnRightEdge property
It was implicitly checking for the value of item.containment before
ensuring that it existed, generating a warning in the log on every
plasmashell load.

Guard against that and also optimize for the case where sidePanelStack
isn't even active yet. In the process, make it const and factor in a
second property that isn't needed anymore.