applets/window-list: use new popup placement enum
The window-list applet still uses the old location for the popup
placement enum (in PlasmaCore.Types), rather than its new home in
PlasmaExtras.Menu. As the old one is no longer defined, this may lead to
a misaligned popup menu, and definitely to a runtime error.

This change adapts the applet to the new way of accessing the popup
placement types.