applets/kickoff: Move enter key handling to listviews
Kickoff has a model where the main view forwards key events to the
search field, and the search field forwards keys to the main view.

Right now the enter key is handled by the search field to activate
the current item in the view.

This doesn't make sense semantically: committing input text should
not activate an item, it should apply the search, which isn't needed
as we search as we type. This could break input methods and the
clear button.

Activating the current item in a view should be encapsulated within
the view.

BUG: 479233
FIXED-IN: 6.2.0