Make clicking terminology more consistent
Throughout plasma, "(left/middle/right/double)-click" is usually
spelled with a hyphen, but not always; sometimes an un-hyphenated
spelling is used. Sometimes a kcm or other dialog will even mix both
spellings. Yet another form is "left button click"; distinguishing
actual button presses from semantic actions makes sense with clicking
emulations etc., but this use is not consistent. In addition, the
touchpad kcm is not consistent about the use of clicking, tapping and
pressing.

This change adds the hyphen to the uses of clicks and clicking actions,
and separates clicking physical actions (e.g. "Press left and right
buttons for middle-click"). Interacting with a touchpad is now called
pressing when it involves the integrated button, never tapping.
Additional labels are added to the radio buttons for the integrated
right click menu indicating how to middle-click by pressing (only
shown if the hardware does not support showing the separate middle-click
configuration. The affected strings (and their control groups, if
applicable) also have their translation contexts added or made more
informative, and minor grammatical errors are fixed ("tap with two
finger").