applets/showActivityManager: Only show activityManager when drag contains windows
It doesn't make sense to show it on something that can't be dropped
into it.
This fixes the crash in the mentioned bug but there is a more general
issue that applets can interfere with the edit mode widget dragging.
BUG:492624
FIXED-IN:6.1.5


(cherry picked from commit d4b87ac6f4d2c323be149c03a61959fc74d6435c)

Co-authored-by: David Redondo <kde@david-redondo.de>