Add custom prefix to user generated desktop files made from shortcuts KCM
The shortcuts KCM re-uses desktop files to store launchers for custom
commands. This is extremely sensible design as it is something being
launched and the design we want. Semantically we're creating a new entry
point. It's the only way you'll get the right cgroup mapped on launch. 

However the name entered by the user could clash with something on the
system which is something we don't want. Most apps are prefixed with the
correct reverse DNS prefix so it's a rare case, but not all.

This was seen in the wild by a user overriding alacritty.desktop by
accident.

Adding a custom prefix for user-generated files should avoid any
potential clash.

Fixed-in: 6.2
BUG: 492754