desktoppackage: fix title case in widget explorer
Widget explorer has a button that confirms uninstalling widgets. This
button uses title case as recommended by the HIG, but does so
incorrectly; it capitalizes the word "of" but short prepositions of less
than four characters are supposed to be lower case in APA style (and
most other style guides).

This change makes this label use correct title case and adds translation
context.

BUG: 492790
FIXED-IN: 6.3